requirements.txt
fastapi
uvicorn[standard]
matplotlib
numpy
requests
python-dotenv
PySimpleGUI --extra-index-url https://PySimpleGUI.net/install
